This modification contains the new version of the Relational Algebra.
There are two new algebras called Relation-C++ and ExtRelation-C++ which
corresponds to the new Relational Algebra and the new Extended Relational
Algebra. The Relational Algebra contains the type constructors "rel" and
"tuple" and the main operations like feed, consume, project, product, rename,
filter, and count. The Extended Relational Algebra has the extended
operations such as sum, avg, loopjoin, hashjoin, sort, sortby, etc. For more
information on these algebras one can try the pd-system documentation.

  The Relational Algebra supports both the Main Memory and Persistent models.
There is a define called "PERSISTENT_RELALG" that can be switched on and off
for this purposes. This define can be seen in the makefiles of the Relational
Algebra and the Extended Relational Algebra and must be set together. One
should remember that the databases created in these two models are not
compatible. I expect that the databases created with the Old Main Memory
Relational Algebra are compatible with the new (Main Memory) one, but I did
not test it. In case of problems, all databases must be re-created.

  The Old Main Memory Relational Algebra is kept alive. There is an algebra
called OldRelational-C++ that corresponds to it. The type constructors were
renamed to "mrel" and "mtuple" and the operators still have the same name.

  The Array Algebra points to the Old Relational Algebra, because it would be
easier to the algebra implementor to make the conversion.

  All type constructors that will be part of a relation must have a memory
representation without pointers. This need was discussed in the Secondo
Programmers Guide. In this way, the FText Algebra will not work inside
relations, as it is implemented now. As a suggestion, one could use FLOBs to
implement the "text" inside the FText Algebra.