Several changes have been done in this version, but the most important ones are:
- The Get method of the FLOB class (also DBArray) does not copy the values
to the result anymore, it just passes a const pointer. A const pointer is
necessary because the FLOB can be cached and I want to avoid changes in
its memory representation. All algebras that handle FLOBs are already
modified to cope with this new feature.

- Tuples contain a tuple type (TupleType) to store their metadata. At
every tuple creation, a copy of the tuple type was done. In
this version this copy is avoided, several tuples share the
same tuple type, keeping only a pointer to it. A reference counter is
added in the TupleType class.

- Every algebra keeps a pointer to the global instances of NestedList and
QueryProcessor. I added a pointer to the AlgebraManager too, because
calling the function to retrieve it inside tuples was consuming too much
time.

- It is now possible to restore databases in the old format, i.e.
before my last big commit at 09.01.2006.