type templates INDEX_TYPE and RESULT_TYPE are using unsigned int now
If the operator new isn't able to allocate the requested amount of memory, the program terminates. Due to this functionality Secondo would become instable.

That's why the operator new was rolled back to function calloc.

The effect could be shown by the following example:
   #include <stdio.h>
   int main(void) {
      int *x;
      x=new int[1000000000];
      fprintf(stderr, "%X\n", x);
      return 0;
   }
The binary returns the following error message:
   terminate called after throwing an instance of 'std::bad_alloc'
     what():  St9bad_alloc
   Aborted