Prevent the decrease of the stored file id.
This can happen when another SECONDO instance
requests a new file id, and we perform a scan before
the file is created in the file system. In this
case, we would overwrite the value with an outdated
file id.

In addition, the get operation aquires a write lock
to prevent deadlocks in the read-modify-write cycle.