Removed custom DEFAULT_LISTEN_QUEUE_SIZE setting
So far, the DEFAULT_LISTEN_QUEUE_SIZE was used to control the
backlog size of the sockets. This value was set to a low
value of 5, which can lead to refused connections on
busy systems.

However, the recommended queue size has been 128 for 20 years,
and newer Linux systems recommend a value of 4096. This patch
removes the custom DEFAULT_LISTEN_QUEUE_SIZE and uses the
SOMAXCONN value of the used operating system (e.g., 128 on
Ubuntu 18.04 and 4096 on Ubuntu 20.04).